removed warnings
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@6156 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/netimport/NINavTeqHelper.cpp
@@ -166,7 +166,7 @@ NINavTeqHelper::addVehicleClasses(NBEdge &e, const std::string &oclassS) throw()
 void
 NINavTeqHelper::addVehicleClass(NBEdge &e, SUMOVehicleClass c) throw()
 {
-    for (size_t i=0; i<e.getNoLanes(); i++) {
+    for (unsigned int i=0; i<e.getNoLanes(); i++) {
         e.allowVehicleClass(i, c);
     }
 }
